@@ -12,7 +12,7 @@
   </PropertyGroup>
 
   <PropertyGroup Label="General Settings">
-    <LangVersion>11.0</LangVersion>
+    <LangVersion>12.0</LangVersion>
     <DebugType>embedded</DebugType>
     <DebugSymbols>true</DebugSymbols>
     <EmbedUntrackedSources>true</EmbedUntrackedSources>
