@@ -1,14 +1,17 @@
-﻿using System;
-using System.Collections.ObjectModel;
+﻿using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Text;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
+using Community.VisualStudio.Toolkit;
 using DataGridExtensions;
 using Microsoft.Build.Construction;
 using Microsoft.Build.Evaluation;
+using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Shell.Interop;
+using NuGet.Packaging.Core;
 using NuGet.Versioning;
 using NuGetMonitor.Abstractions;
 using NuGetMonitor.Model.Models;
@@ -21,9 +24,10 @@ using TomsToolbox.Wpf;
 namespace NuGetMonitor.View.Monitor;
 
 #pragma warning disable CA1812 // Avoid uninstantiated internal classes => used in xaml!
+
 internal sealed partial class NuGetMonitorViewModel : INotifyPropertyChanged
 {
-    private static readonly string[] _versionMetadataNames = { "Version", "VersionOverride" };
+    private static readonly string[] _versionMetadataNames = ["Version", "VersionOverride"];
 
     private readonly ISolutionService _solutionService;
 
@@ -79,7 +83,7 @@ internal sealed partial class NuGetMonitorViewModel : INotifyPropertyChanged
 
             var packages = packageReferences
                 .GroupBy(item => item.Identity)
-                .Select(group => new PackageViewModel(group, PackageItemType.PackageReference, _solutionService))
+                .Select(group => new PackageViewModel(this, group, PackageItemType.PackageReference, _solutionService))
                 .ToArray();
 
             var packageIds = packages
@@ -93,7 +97,7 @@ internal sealed partial class NuGetMonitorViewModel : INotifyPropertyChanged
                 .SelectMany(project => project.Project.CentralVersionMap.Values.Select(item => new PackageReferenceEntry(item.EvaluatedInclude, item.GetVersion() ?? VersionRange.None, item, project, item.GetMetadataValue("Justification"), false)))
                 .Where(item => !packageIds.Contains(item.Identity.Id))
                 .GroupBy(item => item.Identity)
-                .Select(item => new PackageViewModel(item, PackageItemType.PackageVersion, _solutionService))
+                .Select(item => new PackageViewModel(this, item, PackageItemType.PackageVersion, _solutionService))
                 .ToArray();
 
             Packages = packages.Concat(transitivePins).ToArray();
@@ -129,7 +133,7 @@ internal sealed partial class NuGetMonitorViewModel : INotifyPropertyChanged
         Load();
     }
 
-    public static void Update(PackageViewModel packageViewModel)
+    public void Update(PackageViewModel packageViewModel)
     {
         Update(new[] { packageViewModel });
     }
@@ -139,60 +143,112 @@ internal sealed partial class NuGetMonitorViewModel : INotifyPropertyChanged
         Update(SelectedPackages.ToArray());
     }
 
-    private static void Update(ICollection<PackageViewModel> packageViewModels)
+    private async void Update(ICollection<PackageViewModel> packageViewModels)
     {
-        using var projectCollection = new ProjectCollection();
-
-        var packageReferencesByProject = packageViewModels
-            .Where(viewModel => viewModel.IsUpdateAvailable)
-            .SelectMany(viewModel => viewModel.Items.Select(item => new { item.Identity.Id, item.VersionSource, viewModel.ActiveVersion, item.VersionSource.GetContainingProject().FullPath, viewModel.SelectedVersion }))
-            .GroupBy(item => item.FullPath);
-
-        foreach (var packageReferenceEntries in packageReferencesByProject)
+        try
         {
-            var fullPath = packageReferenceEntries.Key;
+            IsLoading = true;
 
-            var project = ProjectRootElement.Open(fullPath, projectCollection, true);
+            using var projectCollection = new ProjectCollection();
 
-            var projectItems = project.Items;
+            var viewModels = packageViewModels
+                .Where(viewModel => viewModel.IsUpdateAvailable && viewModel.SelectedVersion is not null)
+                .ToArray();
 
-            foreach (var packageReferenceEntry in packageReferenceEntries)
+            if (!await AreTargetFrameworksCompatible(viewModels))
+                return;
+
+            var packageReferencesByProject = viewModels
+                .SelectMany(viewModel => viewModel.Items.Select(item => new { item.Identity.Id, item.VersionSource, viewModel.ActiveVersion, item.VersionSource.GetContainingProject().FullPath, viewModel.SelectedVersion }))
+                .GroupBy(item => item.FullPath);
+
+            foreach (var packageReferenceEntries in packageReferencesByProject)
             {
-                var id = packageReferenceEntry.Id;
-                var selectedVersion = packageReferenceEntry.SelectedVersion;
-                var versionSource = packageReferenceEntry.VersionSource;
-                var currentVersion = packageReferenceEntry.ActiveVersion switch
+                var fullPath = packageReferenceEntries.Key;
+
+                var project = ProjectRootElement.Open(fullPath, projectCollection, true);
+
+                var projectItems = project.Items;
+
+                foreach (var packageReferenceEntry in packageReferenceEntries)
                 {
-                    NuGetVersion version => version.OriginalVersion,
-                    VersionRange versionRange => versionRange.OriginalString,
-                    _ => null
-                };
+                    var id = packageReferenceEntry.Id;
+                    var selectedVersion = packageReferenceEntry.SelectedVersion;
+                    var versionSource = packageReferenceEntry.VersionSource;
+                    var currentVersion = packageReferenceEntry.ActiveVersion switch
+                    {
+                        NuGetVersion version => version.OriginalVersion,
+                        VersionRange versionRange => versionRange.OriginalString,
+                        _ => null
+                    };
 
-                if (selectedVersion == null || currentVersion == null)
-                    continue;
+                    if (selectedVersion == null || currentVersion == null)
+                        continue;
 
-                var metadataItems = projectItems
-                    .Where(item => item.ItemType == versionSource.ItemType)
-                    .Where(item => string.Equals(item.Include, id, StringComparison.OrdinalIgnoreCase))
-                    .Select(item => item.Metadata.FirstOrDefault(metadata => _versionMetadataNames.Any(name => string.Equals(metadata.Name, name, StringComparison.OrdinalIgnoreCase))
-                                                                          && string.Equals(metadata.Value, currentVersion, StringComparison.OrdinalIgnoreCase)))
-                    .ExceptNullItems();
+                    var metadataItems = projectItems
+                        .Where(item => item.ItemType == versionSource.ItemType)
+                        .Where(item => string.Equals(item.Include, id, StringComparison.OrdinalIgnoreCase))
+                        .Select(item => item.Metadata.FirstOrDefault(metadata => _versionMetadataNames.Any(name => string.Equals(metadata.Name, name, StringComparison.OrdinalIgnoreCase))
+                                                                                 && string.Equals(metadata.Value, currentVersion, StringComparison.OrdinalIgnoreCase)))
+                        .ExceptNullItems();
 
-                foreach (var metadata in metadataItems)
-                {
-                    metadata.Value = selectedVersion.ToString();
+                    foreach (var metadata in metadataItems)
+                    {
+                        metadata.Value = selectedVersion.ToString();
+                    }
                 }
+
+                project.Save();
             }
 
-            project.Save();
-        }
+            foreach (var packageViewModel in viewModels)
+            {
+                packageViewModel.ApplySelectedVersion();
+            }
 
-        foreach (var packageViewModel in packageViewModels)
+            ProjectService.ClearCache();
+        }
+        catch (OperationCanceledException)
         {
-            packageViewModel.ApplySelectedVersion();
+            // session cancelled
+        }
+        catch (Exception ex)
+        {
+            Log(LogLevel.Error, $"Updating package failed: {ex}");
+        }
+        finally
+        {
+            IsLoading = false;
+        }
+    }
+
+    private static async Task<bool> AreTargetFrameworksCompatible(PackageViewModel[] viewModels)
+    {
+        foreach (var viewModel in viewModels)
+        {
+            var packageDetails = await NuGetService.GetPackageDetails(new PackageIdentity(viewModel.PackageReference.Id, viewModel.SelectedVersion));
+            if (packageDetails == null)
+                continue;
+
+            var projects = viewModel.Items.Select(referenceEntry => referenceEntry.ProjectItemInTargetFramework.Project);
+
+            var incompatibleProjects = projects
+                .Where(project => !packageDetails.SupportedFrameworks.Any(supportedFramework => project.TargetFramework.IsCompatibleWith(supportedFramework)))
+                .ToArray();
+
+            if (incompatibleProjects.Length <= 0)
+                continue;
+
+            var response = await VS.MessageBox.ShowAsync(
+                $"Package {packageDetails.Identity} ({string.Join(", ", packageDetails.SupportedFrameworks)}) is not compatible with projects {string.Join(",", incompatibleProjects.Select(p => p.NameAndFramework))}",
+                "Do you want to update anyway?",
+                OLEMSGICON.OLEMSGICON_WARNING, OLEMSGBUTTON.OLEMSGBUTTON_YESNO, OLEMSGDEFBUTTON.OLEMSGDEFBUTTON_SECOND);
+
+            if (response != VSConstants.MessageBoxResult.IDYES)
+                return false;
         }
 
-        ProjectService.ClearCache();
+        return true;
     }
 
     private void NormalizePackageReferences()
